pyproject.toml
[tool.poetry]
name = "social-network-startnavi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "social_network_startnavi"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
djangorestframework = "^3.14.0"
gunicorn = "^20.1.0"
django-jazzmin = "^2.6.0"
psycopg2-binary = "^2.9.6"
django-split-settings = "^1.2.0"
django-cors-headers = "^3.14.0"
django-autocomplete-light = "^3.9.4"
drf-spectacular = "^0.26.1"
djangorestframework-simplejwt = "^5.2.2"
celery = "^5.2.7"
amqp = "^5.1.1"
django-extensions = "^3.2.1"
redis = "^4.5.4"
ipython = "^8.12.0"
requests = "^2.28.2"
faker = "^18.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
